requirements.txt
absl-py==0.12.0
appdirs==1.4.4
asgiref==3.2.10
astor==0.8.1
astunparse==1.6.3
beautifulsoup4==4.9.2
bs4==0.0.1
cachetools==4.2.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
distlib==0.3.1
dj-database-url==0.5.0
dj.choices==0.11.0
Django==3.1
django-cities==0.6
django-countries==7.1
django-heroku==0.3.1
django-model-utils==4.1.1
django-phonenumber-field==5.1.0
filelock==3.0.12
flatbuffers==1.12
freeze==3.0
gast==0.3.3
google-auth==1.28.0
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.7
importlib-metadata==3.7.3
importlib-resources==5.1.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.4
mock==4.0.3
mysqlclient==2.0.1
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
phonenumbers==8.12.21
Pillow==7.2.0
protobuf==3.15.6
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
pyuca==1.2
qrcode==6.1
requests==2.19.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.5.4
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
swapper==1.1.2.post1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
tqdm==4.60.0
typing-extensions==3.7.4.3
urllib3==1.23
virtualenv==20.4.3
virtualenvwrapper-win==1.2.6
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
zipp==3.4.1
